+++ packages/orq-rc/src/orq_ai_sdk/models/createcontactop.py
@@ -140,7 +140,7 @@ class CreateContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-23T13:46:12.112Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-24T04:49:59.822Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")