+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatapointop.py
@@ -373,17 +373,27 @@ r"""The type of the content part. Always `file`."""
 
 
 class RetrieveDatapoint2FileTypedDict(TypedDict):
-    file_data: str
+    file_data: NotRequired[str]
     r"""The file data as a data URI string in the format 'data:<mime-type>;base64,<base64-encoded-data>'. Example: 'data:image/png;base64,iVBORw0KGgoAAAANS...'"""
-    filename: str
+    uri: NotRequired[str]
+    r"""URL to the file. Only supported by Anthropic Claude models for PDF files."""
+    mime_type: NotRequired[str]
+    r"""MIME type of the file (e.g., application/pdf, image/png)"""
+    filename: NotRequired[str]
     r"""The name of the file, used when passing the file to the model as a string."""
 
 
 class RetrieveDatapoint2File(BaseModel):
-    file_data: str
+    file_data: Optional[str] = None
     r"""The file data as a data URI string in the format 'data:<mime-type>;base64,<base64-encoded-data>'. Example: 'data:image/png;base64,iVBORw0KGgoAAAANS...'"""
 
-    filename: str
+    uri: Optional[str] = None
+    r"""URL to the file. Only supported by Anthropic Claude models for PDF files."""
+
+    mime_type: Annotated[Optional[str], pydantic.Field(alias="mimeType")] = None
+    r"""MIME type of the file (e.g., application/pdf, image/png)"""
+
+    filename: Optional[str] = None
     r"""The name of the file, used when passing the file to the model as a string."""
 
 
@@ -662,7 +672,7 @@ class RetrieveDatapointEvaluations3(BaseModel):
 
     source: Optional[RetrieveDatapointEvaluationsDatasetsResponseSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2025-10-23T13:46:21.432Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2025-10-24T04:50:09.596Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -714,7 +724,7 @@ class RetrieveDatapointEvaluations2(BaseModel):
 
     source: Optional[RetrieveDatapointEvaluationsDatasetsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2025-10-23T13:46:21.432Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2025-10-24T04:50:09.596Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -766,7 +776,7 @@ class RetrieveDatapointEvaluations1(BaseModel):
 
     source: Optional[RetrieveDatapointEvaluationsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2025-10-23T13:46:21.432Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2025-10-24T04:50:09.596Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -853,5 +863,5 @@ class RetrieveDatapointResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-23T13:46:12.112Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-24T04:49:59.822Z")
     r"""The date and time the resource was last updated"""